Init CUDA kernel

+++ av_hw/cuda/cvt_color.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-void nv12_to_rgb(uint8_t *in_y, uint8_t *in_uv, uint8_t *out_rgb, size_t height, size_t width);
+void nv12_to_rgb(uint8_t *in_y, uint8_t *in_uv, uint8_t *out_rgb, int height, int width);
